# Optimizing softmax & log_softmax