This does similar things as #652.

It additionally re-uses as much computation as possible. See in code comments for how we do this.

Timings on my CPU for `cargo bench --bench softmax`
| branch | fwd | bwd |
| --- | --- | --- |
| main | 1000ms | 1000ms |
| this | ~790ms | ~770ms |